broken down commands
